feat: 分组sql

+++ backend/internal/repo/relationDB/migrations_sqls/uns.sql
@@ -325,6 +325,6 @@ CREATE TABLE if not exists "resource_group" (
 );
 COMMENT ON COLUMN "resource_group"."type" IS '1-sourceFlow 2-eventFlow 3-dashboard';
 
-alter table uns_dashboard add if not exists "group_id" integer null default NULL;
-alter table supos_node_flows add if not exists "group_id" integer null default NULL;
-alter table supos_event_flows add if not exists "group_id" integer null default NULL;
+alter table uns_dashboard add if not exists "group_id" int8 null default NULL;
+alter table supos_node_flows add if not exists "group_id" int8 null default NULL;
+alter table supos_event_flows add if not exists "group_id" int8 null default NULL;
